Incorporate API
Added an API call to the "/" and to the "/search" route.
Left console.logs in to demenstrate results from the API calls. Have commented out the res.json calls that should be used to send the needed data to the front end.
Added the "/search" route just as an added capability of the site. It takes a request from the client containing the 'ticker' key and quaries the API for the full name, market open price and profit margin. Hardcoded to "AMRN" right now for demenstration purposes. Feel free to delete if uneeded.